+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # HidingBar - Burning Crusade
 
-* Fixed missing icon when updating Masque style.
-* Fixed OnEnter handler.
+* The interface has been changed.
+* Added a system of profiles.
+* Added a system of bars.
+* Added "Clip button" option.